+++ src/lib.rs
@@ -52,6 +52,7 @@ pub mod error;
 #[cfg(feature = "graph")]
 pub mod graph;
 pub mod io;
+pub mod memory_ctl;
 pub mod mq;
 pub mod persistence;
 pub mod protocol;